Move ECLIPSE I/O Into Main OPM Common Library
This makes the facility usable for the restart read/write code.
