+++ tests/test_EclFilesComparator.cpp
@@ -21,6 +21,7 @@
 #define BOOST_TEST_MODULE EclFilesComparatorTest
 
 #include <boost/test/unit_test.hpp>
+
 #include <examples/test_util/EclFilesComparator.hpp>
 
 BOOST_AUTO_TEST_CASE(deviation) {
